backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
motor>=3.3.2
pymongo>=4.6.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.6

# Environment and Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic[email]>=2.5.0

# AI and Machine Learning (Optional - install if needed)
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=1.5.3

# HTTP Client
httpx>=0.25.2

# Logging and Monitoring
structlog>=23.2.0

# Development and Testing (Optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1

# File handling (Optional)
Pillow>=10.1.0

# CORS handling
fastapi-cors>=0.0.6

# Additional dependencies for crypto
cryptography>=3.4.8
ecdsa>=0.18.0
rsa>=4.9.0
pyasn1>=0.4.8